+++ morpheus-graphql-examples-scotty/src/Server/Sophisticated/API.hs
@@ -14,7 +14,7 @@
 module Server.Sophisticated.API
   ( api,
     EVENT,
-    gqlRoot,
+    root,
   )
 where
 
@@ -25,8 +25,8 @@ import qualified Data.Map as M
   )
 -- MORPHEUS
 import Data.Morpheus
-  ( debugInterpreter,
-    interpreter,
+  ( deriveApp,
+    runApp,
   )
 import Data.Morpheus.Document
   ( importGQLDocumentWithNamespace,
@@ -41,7 +41,6 @@ import Data.Morpheus.Types
     ID,
     Input,
     MUTATION,
-    QUERY,
     Resolver,
     ResolverM,
     ResolverQ,
@@ -56,7 +55,6 @@ import Data.Morpheus.Types
     publish,
     subscribe,
   )
-import Data.Proxy (Proxy (..))
 import Data.Set (Set)
 import qualified Data.Set as S
   ( fromList,
@@ -117,10 +115,10 @@ newtype Content = Content {contentID :: Int}
 type EVENT = Event Channel Content
 
 api :: Input api -> Stream api EVENT IO
-api = debugInterpreter gqlRoot
+api = runApp (deriveApp root)
 
-gqlRoot :: RootResolver IO EVENT Query Mutation Subscription
-gqlRoot =
+root :: RootResolver IO EVENT Query Mutation Subscription
+root =
   RootResolver
     { queryResolver,
       mutationResolver,