+++ src/Data/Morpheus.hs
@@ -1,9 +1,41 @@
+{-# LANGUAGE FlexibleContexts #-}
+
 -- | Build GraphQL APIs with your favourite functional language!
 module Data.Morpheus
-  ( Interpreter (..),
+  ( interpreter,
+    debugInterpreter,
+    App,
+    AppRunner (..),
+    deriveApp,
   )
 where
 
-import Data.Morpheus.Server.Deriving.Interpreter
-  ( Interpreter (..),
+-- MORPHEUS
+import Data.Morpheus.Core
+  ( App,
+    AppRunner (..),
   )
+import Data.Morpheus.Server.Deriving.App
+  ( RootResolverConstraint,
+    deriveApp,
+  )
+import Data.Morpheus.Types
+  ( RootResolver (..),
+  )
+
+-- | main query processor and resolver
+interpreter ::
+  AppRunner e m a b =>
+  (RootResolverConstraint m e query mut sub) =>
+  RootResolver m e query mut sub ->
+  a ->
+  b
+interpreter = runApp . deriveApp
+
+debugInterpreter ::
+  AppRunner e m a b =>
+  (RootResolverConstraint m e query mut sub) =>
+  RootResolver m e query mut sub ->
+  a ->
+  b
+debugInterpreter = debugApp . deriveApp