+++ src/Data/Morpheus/Document.hs
@@ -16,7 +16,7 @@ import Data.ByteString.Lazy.Char8
 import Data.Morpheus.Core
   ( render,
   )
-import Data.Morpheus.Server.Deriving.Resolve
+import Data.Morpheus.Server.Deriving.App
   ( RootResolverConstraint,
     deriveSchema,
   )